# bridge/bridge/api/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bridge API Documentation
  description: |
    API for handling token wrapping and unwrapping between native and wrapped networks.

    The API exposes internal bridge state so external systems can monitor request processing,
    payout status, errors, and fee/conversion estimates.
    The API returns all requests and errors recorded since the bridge began operating.

    Unwrap operations are supported only in wrap mode and stake mode (not available in swap mode).

    Please check the Bridge README:
    <https://github.com/symbol/product/blob/dev/bridge/README.md>
  version: 1.0.0
servers:
  - url: "../../"
    description: Current server
  - url: "http://localhost:5000"
    description: Local development server
...
paths:
...
  /wrap/requests:
    get:
      summary: Returns wrap requests information.
      description: |
        Retrieves a paginated list of all wrap requests detected since the start of the bridge operation on the native network.
        These represent incoming transactions that initiate wrapping.
        If a request results in an error, it will not appear in `/wrap/requests`; use `/wrap/errors` to retrieve it.
      tags:
        - Wrap
      operationId: getWrapRequests
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/PayoutStatus'
      responses:
        200:
          description: List of wrap requests recorded by the bridge.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WrapRequest'
        400:
          description: Invalid request parameters (e.g., invalid pagination or filters).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wrap/requests/{address}:
    get:
      summary: Returns wrap requests filtered by sender or destination address.
      description: |
        Retrieves a paginated list of all wrap requests associated with the provided address since the start of the bridge operation.

        The address filter matches either:
          - the **request (sender) address on the native network**, or
          - the **destination address on the wrapped network**.

        Addresses from either network are valid and will be used to filter wrap requests accordingly.
      tags:
        - Wrap
      operationId: getWrapRequestsByAddress
      parameters:
        - $ref: '#/components/parameters/Address'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/PayoutStatus'
      responses:
        200:
          description: Wrap requests linked to the specified address.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WrapRequest'
        400:
          description: Invalid request parameters (e.g., invalid pagination, filters, or address).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wrap/requests/hash/{transaction_hash}:
    get:
      summary: Returns wrap requests by request or payout transaction hash.
      description: |
        Retrieves a paginated list of all wrap requests matched by either the request hash or the associated payout hash since the start of the bridge operation.
        
        The **request hash** refers to the transaction hash of the originating transaction
        on the **native network**, while the **payout hash** refers to the transaction hash
        of the corresponding payout transaction on the **wrapped network**.

        The hash must be a valid 64-character hexadecimal string.
        **Important:** For Ethereum, the hash MUST be provided **without** the `0x` prefix.
      tags:
        - Wrap
      operationId: getWrapRequestsByHash
      parameters:
        - $ref: '#/components/parameters/TransactionHash'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/PayoutStatus'
      responses:
        200:
          description: Wrap requests corresponding to the specified transaction hash.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WrapRequest'
        400:
          description: Invalid request parameters (e.g., invalid pagination, filters, or transaction hash).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wrap/errors:
    get:
      summary: Returns wrap errors.
      description: |
        Retrieves a paginated list of all errors encountered since the start of the bridge operation while processing wrap requests.
        If a request results in an error, it will not appear in `/wrap/requests`; use this endpoint to retrieve it.
      tags:
        - Wrap
      operationId: getWrapErrors
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        200:
          description: List of wrap errors stored by the bridge.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WrapError'
        400:
          description: Invalid request parameters (e.g., invalid pagination).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wrap/errors/{address}:
    get:
      summary: Returns wrap errors for a specific sender address.
      description: |
        Retrieves a paginated list of all wrap errors associated with the provided address since the start of the bridge operation.
        The address filter matches the **request address on the native network** (sender).
        Addresses from the native network are valid and will be used to filter wrap request errors accordingly.
        Filtering by destination address is not supported because destination information is not available for errored requests.
      tags:
        - Wrap
      operationId: getWrapErrorsByAddress
      parameters:
        - $ref: '#/components/parameters/Address'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        200:
          description: List of wrapping errors associated with the provided address.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WrapError'
        400:
          description: Invalid request parameters (e.g., invalid pagination or address).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wrap/errors/hash/{transaction_hash}:
    get:
      summary: Returns wrap errors for a specific transaction hash.
      description: |
        Retrieves a paginated list of all wrap errors matched by the request transaction hash since the start of the bridge operation.

        The **request hash** refers to the transaction hash of the originating transaction on the **native network**.
        The hash must be a valid 64-character hexadecimal string.

        **Important:** For Ethereum, the hash MUST be provided **without** the `0x` prefix.
      tags:
        - Wrap
      operationId: getWrapErrorsByHash
      parameters:
        - $ref: '#/components/parameters/TransactionHash'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        200:
          description: List of wrapping errors associated with the provided transaction hash.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WrapError'
        400:
          description: Invalid request parameters (e.g., invalid pagination or transaction hash).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wrap/prepare:
    post:
      summary: Estimates the cost of a wrap transaction.
      description: |
        Estimates the outcome of a wrap operation, including conversion results and fees.
        This endpoint simulates how many wrapped tokens the user will receive based on the bridge current configuration, fee structure, and mode.
        If the estimated amount is acceptable, submit the actual wrap request by transferring the desired amount to the bridge account as explained in the README.
      tags:
        - Wrap
      operationId: prepareWrap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrepareWrapRequest'
      responses:
        200:
          description: Estimated wrap transaction details, including fees and net received amount.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrepareWrapResponse'
        400:
          description: Wrap request parameters are invalid or violate bridge limits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal bridge error during preparation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /unwrap/requests:
    get:
      summary: Returns unwrap requests information.
      description: |
        Retrieves a paginated list of all unwrap requests submitted since the start of the bridge operation on the wrapped network.
        These represent incoming transactions that initiate unwrapping.
        If a request results in an error, it will not appear in `/unwrap/requests`; use `/unwrap/errors` to retrieve it.
      tags:
        - Unwrap
      operationId: getUnwrapRequests
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/PayoutStatus'
      responses:
        200:
          description: List of unwrap requests recorded by the bridge.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnwrapRequest'
        400:
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /unwrap/requests/{address}:
    get:
      summary: Returns unwrap requests filtered by sender or destination address.
      description: |
        Retrieves a paginated list of all unwrap requests associated with the provided address since the start of the bridge operation.

        The address filter matches either:
          - the **request (sender) address on the wrapped network**, or
          - the **destination address on the native network**.

        Addresses from either network are valid and will be used to filter unwrap requests accordingly.
      tags:
        - Unwrap
      operationId: getUnwrapRequestsByAddress
      parameters:
        - $ref: '#/components/parameters/Address'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/PayoutStatus'
      responses:
        200:
          description: Unwrap requests linked to the specified address.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnwrapRequest'
        400:
          description: Invalid request parameters (e.g., invalid pagination, filters, or address).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /unwrap/requests/hash/{transaction_hash}:
    get:
      summary: Returns unwrap requests for a transaction hash.
      description: |
        Retrieves a paginated list of all unwrap requests matched by either the request hash or the associated payout hash since the start of the bridge operation.

        The **request hash** refers to the transaction hash of the originating transaction on the **wrapped network**,
        while the **payout hash** refers to the transaction hash of the corresponding payout transaction on the **native network**.

        The hash must be a valid 64-character hexadecimal string.
        **Important:** For Ethereum, the hash MUST be provided **without** the `0x` prefix.
      tags:
        - Unwrap
      operationId: getUnwrapRequestsByHash
      parameters:
        - $ref: '#/components/parameters/TransactionHash'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/PayoutStatus'
      responses:
        200:
          description: Unwrap requests corresponding to the given transaction hash.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnwrapRequest'
        400:
          description: Invalid request parameters (e.g., invalid pagination, filters, or transaction hash).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /unwrap/errors:
    get:
      summary: Returns unwrap errors.
      description: |
        Retrieves a paginated list of all errors encountered since the start of the bridge operation while processing unwrap requests.
        If a request results in an error, it will not appear in `/unwrap/requests`; use this endpoint to retrieve it.
      tags:
        - Unwrap
      operationId: getUnwrapErrors
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        200:
          description: List of unwrap-related errors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnwrapError'
        400:
          description: Invalid request parameters (e.g., invalid pagination, filters).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /unwrap/errors/{address}:
    get:
      summary: Returns unwrap errors for a specific address.
      description: |
        Retrieves a paginated list of all unwrap errors associated with the provided address since the start of the bridge operation.

        The address filter matches the **request address on the wrapped network** (sender). 
        Addresses from the wrapped network are valid and will be used to filter unwrap request errors accordingly.
        
        Filtering by destination address is not supported because destination information is not available for errored requests.
      tags:
        - Unwrap
      operationId: getUnwrapErrorsByAddress
      parameters:
        - $ref: '#/components/parameters/Address'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        200:
          description: Unwrap errors associated with the given address.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnwrapError'
        400:
          description: Invalid request parameters (e.g., invalid pagination, filters, or address).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /unwrap/errors/hash/{transaction_hash}:
    get:
      summary: Returns unwrap errors for a specific transaction hash.
      description: |
        Retrieves a paginated list of all unwrap errors matched by the request transaction hash since the start of the bridge operation.

        The **request hash** refers to the transaction hash of the originating transaction on the **wrapped network**.
        The hash must be a valid 64-character hexadecimal string.

        **Important:** For Ethereum, the hash MUST be provided **without** the `0x` prefix.
      tags:
        - Unwrap
      operationId: getUnwrapErrorsByHash
      parameters:
        - $ref: '#/components/parameters/TransactionHash'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        200:
          description: Unwrap errors associated with the specified transaction hash.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnwrapError'
        400:
          description: Invalid request parameters (e.g., invalid pagination, filters, or transaction hash).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip for pagination.
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 0
        default: 25
    Sort:
      name: sort
      in: query
      required: false
      description: |
        Results are sorted by the block height of the transaction that initiated the request.
        The block height used for sorting is the request originating network block height
        (native network for wrap requests; wrapped network for unwrap requests).
        This field selects the sort direction: `1` (ascending) or `0` (descending).
      schema:
        type: integer
        enum:
          - 1
          - 0
        default: 1
    PayoutStatus:
      name: payout_status
      in: query
      required: false
      description: "Filter by payout processing status: `0` (unprocessed), `1` (sent), `2` (completed), or `3` (failed)."
      schema:
        type: integer
        enum:
          - 0
          - 1
          - 2
          - 3
...